[s08] Made chat3 the default; chat4 currently "experimental"
In the last commit a made a new version of chat - chatv4 - the default
without saying anything about it. This was cause confusion because chips
that used to work will stop working.

This commit makes chat4 an opt-in change. When I'm ready to make the new
standard, I will have detailed instructions on upgrading.