[s08] Convert to using common flashing code
Since we now have memory regions, it seemed to make sense to break up
the boot/chat/vectors page (the last page of the flash) into regions.

The new common flashing code has good, verbose, helpful support for
regions, so by default we get a nice user interface that tells us
exactly what's happening when we are re-flashing chat (which is a risky
and slightly terrifying business).

Chat v3 and chat v4 capture the current Git commit and put it into the
first 4 bytes of the boot page. We would ideally *not* like to consider
a differing commit to be a verify mismatch, so we put the commit into
its own region, and we make sure that the verify process is smart about
ignoring differences in the commit - ie, it doesn't consider them to be
a verification failure.

Chat v2 does *not* embed a commit; hence, when we are setting up to
compile it, we leave the commit region empty, and start the boot region
at the beginning of the page (instead of 4 bytes in).

All of this is very simple, and works beautifully.