[rebuild] open-file returns a flag instead of aborting on error
I have wanted this for a long time! Previously, the fact that open-file
would abort on error meant it was hard to probe for serially- or
USB-connected target devices.

Now it's easy!

For most of the existing uses of open-file? we now immediately call
?abort, but where it is useful we test the result and try something
else.

The target serial code in particular is now much simpler and better and
additionally tries *two* serial ports: serial-target and serial-target2.
Loading of all host-side chat code is prevented if there are no serial
targets to connect to.

This also makes a much improved "auto-id" story possible. I added some
code to do this for the S08 target, but it had problems. This makes
solving them easy. (Those changes are coming.)