[common] Flashing code was broken by a previous commit!
In the process of trying to fix some problems with chunk size, I deleted
the word "region" (in flash-region) that fetches the start address and
size, so flash-region starts at zero and goes forever.

In this commit, that is fixed. I also decided that setting /chunk to
a default value of 256 so that trying to execute verify before chatting
will no longer give an alarming floating point (?) exception (because of
dividing by zero).

Also, the flash-image code *checks* that the chunk size is not bigger
than the erase size, but instead of fixing it, it complains for you to
fix it. Maybe that's not the right choice...