[msp430] Updated to new stack code and interact model
Unfortunately, chat-v4 is a mess in a bunch of ways, but at the moment
it's not worth replacing.

The "new way" of doing interact sends the PC to execute and a stack
pointer (DP) to the target when running code, and then requests the
stack pointer (DP) afterwards.

The chat-v4 code passes a lot of registers back and forth unnecessarily,
but, again, I'm not sure I want to fix it right now. Instead, we mostly
just throw away what is no longer needed, and pass back and forth what
is necessary for chat to work.

But eventually there should probably be a chat-v5.

At least this proves that the interact model works.