[arm] Rewrote and simplified ST-LINK support
There was a lot of unused and experimental code, and some things were
done in an ugly way.

I also wanted to try to use a minimal set of ST-LINK primitives. Other
than the commands to transition to SWD debug mode, there are
only four primitives that are used:

  dap@   dap!      read or write a single 32-bit word
  read32 write32   read or write a series of 32-bit words

Instead of using the ST-LINK debug commands to read and write the core
registers and to stop and start the core, we instead use dap@ and dap!
to read and write the standard ARM debug registers: DHCSR, DCRSR, and
DCRDR.

I'm much happier with the code now.