[arm] Improved auto-detection of STM32 chips (Discovery/Nucleo boards)
If you connect one of the following STM32 boards:

   F0-Discovery
   F072B-Discovery
   F3-Discovery
   F4-Discovery
   F411E-Discovery

   Nucleo-C031C6
   Nucleo-H503RB
   Nucleo-H533RE
   Nucleo-C071RB

and then start muforth like this:

   ./muforth -f target/ARM/board/stm32-discovery-auto.mu4

the chip on the board will be auto-detected (using ST-LINK) and the
correct files for that chip will be loaded.

I will probably rename the board file to something more like
"stm32-auto" to reflect that it works with some Nucleo boards as well,
but I'm keeping the name for now.