[arm] Update ST-LINK support to work with V3 (eg, Nucleo-H533RE)
It turns out that some small changes were necessary to support
ST-LINK/V3 (mostly the way that the jtag version info is queried), and
bigger changes to work with the STM32H5 family, which has an odd quirk:
there are *two* APs (access ports) in the chips's debug peripheral
- unlike every other STM32 that I have dealt with!

This *may* have to do with their implementation of security and
TrustZone; however, it's not mandated by ARM. In fact, the Cortex-M33
TRM (tech refman) only mentions *one* AP in the standard debug
peripheral.

In any case, on the H5 (and possibly the L5 and U5 as well - the newer
Cortex-M33 chips), AP0 (the "normal" AP that you connect to via ST-LINK)
is *only* connected to the DBGMCU block. All you can read from this is
the chip id - the STM32 "MCUID". If you try to read from memory, or
frome the debug space on the core - eg, to get the CPUID or halt the
core - you get a fault.

You have to switch to AP1 to read anything from the chip!

This change "fixes" the ST-LINK support code to be mindful of this
reality, and to support the newer chips correctly.