[arm] More STM32 flash programming improvements
- flash routines file had mismatched .if/.then and wasn't properly loading
  F0/F3 code

- beginning of a "modularization" of per-family flash code

- changed to standardize on use of the word "page" vs "sector"

Right now all the STM32 flash programming code - the actual ARM assembler
code - is in one file with lots of .if/.then conditional compilation. I'm
thinking of extending the per-family "modularization" to include the
actual code and not just "settings". That would mean slightly less
sharing of code, but in the event that I want to add mass erase, bank
erase, option byte programming, etc it would be much easier to manage
with the code in separate files.