[arm] More STM32 flash programming...
Slight clarification of  erase?  (in mu/target/common/flash.mu4) and how
it works for chip families that use a page index rather address when
erasing the flash.

Also, updated the serial bootloader code to convert addresses to page
indices only for devices that normally use addresses when erasing pages.