[arm] Update some STM32 "application" code to use chip-path init
Instead of checking (via .ifdef) which chip we are using, use chip-path
to "calculate" the path of the init.mu4 file.

This doesn't mean that this code will suddenly work on all STM32
families, however!

The chip families are unfortunately very different in terms of which
timers are implemented, and even occasionally the names of registers
(eg, RCC_APB1ENR vs RCC_APBENR1).

I haven't tried to tackle the cross-family compatibility question yet,
even for the small number of chips that I would like to support.