[arm] Updated STM32H533 init to run the chip at 96M
This involved getting all the settings right for the PLL. I *think*
everything is right, but I haven't checked. What I need to do is to set
up a timer to output a 1 kHz square wave - or maybe instead divide the
clock by 10,000 so I get 9.6 kHz out, which I can measure with my DMM.

But I haven't done this.