[arm] Changed ST-LINK code to use a 16 byte private buffer
This is how it used to be, but in a recent "cleanup" of the code
I decided to use  pad  instead.

The trouble is, code that calls into the st-link code - eg, the AVR and
PIC18 programming code - uses  pad  for its own purposes, and the
st-link code's use of  pad  clobbers that.

Obviously, this rather mysteriously broke working code - until I figured
out that the "innocent" change to use  pad  instead of a private buffer
was to blame.

Now the st-link is both cleaner (from the commit that introduced the use
of  pad ) and more useful (since it again uses a private buffer instead
of potentially stepping on the caller's toes).