[arm] Added generic way to enable gpio port clocks (stm32)
I also changed the init code for each chip to use enable-gpio.

Unfortunately each stm32 chip family enables the clocks to the GPIO
ports in a different way. The registers are different and the positions
of the bits are different.

We isolate these differences into a single word that takes a bitmap of
ports (0 bit is port A, 1 bit is port B, etc) and shifts or mangles it
as necessary and writes it to the correct register.