[arm] Use GPIOx_BSRR for both setting and resetting bits
The F4 family lacks the GPIOx_BRR (bit reset register) registers, so any
STM32 code that wants to be cross-platform has to avoid using them.

I'm not sure why ST added the _BRR registers in the first place, since
the _BSRR allows both atomic bit setting and resetting.

So that's what this change does: everywhere I was using _BSRR for
setting and _BRR for resetting has been changed to use only _BSRR;
writing halfwords to the low half to set, and to the high half to reset.