[arm] Change host-side avr/pic programming code to use dap@/dap!
The host machine, when communicating with an STM32 board being used as
a "programming pod", has to read and write a few 32-bit words to send
commands to, and receive results from, the pod; using dap@ and dap! when
reading or writing a single word is simpler and cleaner than reading or
writing a four byte buffer, which is what this code used to do.