[core] Put comment and conditional compilation words in .lex.
This is a small change that is a *very* big change.

It's something that I have wanted to do for a long time: put all the
comment and conditional compilation words in *one* place - a dictionary
chain called  .lex.  - and then make sure that it gets searched at the
right times.

We do the latter by *anchoring* .forth. , .compiler. , and some of the
target-compiler chains to .lex.

.lex. is created by the C code that initializes the dictionary - it was
easier to do it here than to try to "patch it in" later, in Forth - and
this makes it available starting with the first line of mu/startup.mu4,
which defines the "|" comment delimiter in .lex.

Putting everything related to comments and conditional compilation into
.lex. from the beginning means that I don't have to make (or forget to
make!) copies of those into various dictionary chains. Everything is
simpler and more robust.