[common] Improved flexiblity of the common flashing code
I renamed flash-chunk to erase-program-chunk and made it a defer'd word.
This makes the flashing process more flexible.

The AVR and PIC18 were able to reuse flash-region to program eeprom and
config areas, and the code to program STM32 chips that use non-uniform
page sizes was moved *out* of the common flash code where it used to
reside.