[arm,avr] More STM32 chip families can be used as Atmel programmers
Previously, support was limited to F0 and F3 families.

I added F4, C0, and H5 (specifically: F411, C031/C071, H503/H533).

There is some untested code in this commit, but I added the following:

- hardware SPI support. This is required for the actual chip programming
  interface.

- clock output (8M or 12M, depending on chip family (and what's
  possible). This is useful for recovering and programming AVR chips
whose fuses specify an external clock signal; when none is present the
chip cannot be programmed, and the fuses cannot be changed to specify,
eg, an internal clock!

- I2C support

There was skeletal support for doing "packet style" uart send/recv
communication, but I never used it. It is now behind a "with-uart"
ifdef, and it still only works (if it works at all!) on F0 and F3 chips.