[arm, pic18] Program PIC18-Q parts over SPI with flash-image
Instead of having an entirely separate piece of code - with its own
chunking, reporting, etc - I decided to connect the flash-image code
- normally used for programming and verifying flash over a serial chat
interface - to the STM32 programming "pod" SPI interface.

Like the Atmel AVR and 8051 parts, PIC18's use an SPI-like interface
(it's half-duplex, with both target and host sharing the single data
line) for programming.

By connecting the SPI programming interface (which is implemented on an
STM32 development board) with the interact code (the same way that is
done for serial chat) it makes it possible not only to re-use the normal
flash-image and verify words for programming, but it's also possible to
use du to explore the memory contents of the chip!

The ram and i/o are not readable, however, since there is no way to
access them over the SPI programming interface.

But it's useful (and fun) to look at the flash, config space, eeprom,
baked-in device configuration, etc.