[s08] Fixed errors about clobbering the bootloader when flashing
In a previous commit () I put the chat commit into its own region, to
make verifying more accurate: you no longer get a false negative if only
the commit differs (but the rest of the boot/chat code is identical).

When re-flashing the boot/chat code, you have to be running the chat
code on the chip *in RAM*, otherwise the code will attempt to overwrite
itself and fail.

So there is a check, and the process fails with a warning if it finds
that the chat code is running from flash and not ram.

Unfortunately, the earlier commit that put the chat commit into its own
region caused false positives and errored out when it shouldn't. This is
because the boot region was set up to skip the commit - ie, to be
4 bytes into the last page of flash, where the boot/chat code lives.
Even if no boot code had been loaded, the error would be (falsely)
triggered.

Now, both commit and boot regions start out empty, and both are checked
to see if they have encroached on the boot sector.

This is because the chat2 code didn't define a commit, only the boot
code. Testing both, and having both start out empty, is the easiest
solution.