[C core] On Linux, when no USB available, fail gracefully
Previously, if no USB was available (no kernel support, no module(s)
loaded), usb-find-device would error out saying "No such file or
device".

Now, if no tree of USB devices is available - the C code looks for
/proc/bus/usb/ and /dev/bus/usb/ - it simply returns "no match". This
allows the Forth code to continue to load, perhaps skipping the device
it was trying to load.

I ran across this when trying out some code that uses an STM32 device to
program PIC18 devices. I tried to load the PIC18 code, after rebooting
my Windows machine and restarting WSL, and I was surprised to get an
error - instead of a "not found" - and even *more* surprised to find
that lsusb returned absolutely nothing! Usually it at least mentions the
Linux root hub devices.

I thought that the recent Windows update had broken WSL. But then
I plugged in an stm32, and passed it to WSL (using usbipd-win).
Suddenly, /dev/bus/usb/ existed, lsusb acted normal, and my code loaded
successfully.

Newer versions of WSL are using a version of the 6.6 kernel that are
very modularized. The previous kernels had drivers compiled in.

Hopefully this change will make it less painful for those times when you
are coding, but not connected to any USB devices.