[pic18] Rewrote meta-compiler and modified kernel
I rewrote the meta-compiler from the ground up, and modified the
existing, somewhat sketchy and experimental, kernel as well.

It currently supports the following:

- somewhat clever literal compilation
- tail call conversion
- a reasonable implementation of create, does>, and ;code

It is missing all the control structure words - if/else/then, for/next,
do/loop. These should be easy to add, and I have been experimenting with
implementations.

This meta-compiler has an experimental feature that I want to add to all
the others: when making "bracketed" interpreter changes - asm { } for
creating assembler macros, using [ ] to escape from a compilation to do
some computation, or { } to escape to the host forth to execute
a handful of words - a stack is employed so that one always returns to
the correct mode.

Other meta-compilers can occasionally have surprising and unfortunate
results when changing modes.