[arm,pic18,s08] Updated pic18 programming code - via S08 and STM32
The context: Ten years ago (!!) - in late 2014 - I was able to use an
S08 running some simple custom firmware to program a simple chat program
onto to a PIC18F14K50. I was interested in this partcular chip because
it was cheap (at the time - less so now in 2025) and it had a USB
interface. I was at the time a bit obsessed with putting small simple
USB firmware onto 8-bit microcontrollers - starting (and, as it happens,
ending) with the Motorola/Freescale/NXP JS16.

(Aside: I assumed it would be an easy job to write USB firmware for the
14K50 since, rather curiously, the S08 and PIC18 have essentially the
same USB peripheral (perhaps licensed from a third party?). I never did
it. I should probably revisit this.)

Recently, after successfully using an STM32 Discovery board to program
a newer PIC18 - the PIC18F26Q43 (a *very* nice chip indeed) - I wanted
to revisit the 14K50.

Unfortunately, it seems that vendors are constantly making small and
large changes to how the flash on their chips is programmed. ST is
certainly among the culpable in this regard.

The 14K50 - and most of the older PIC18F chips - use a completely
different protocol to the one used on the PIC18F-Q family. From the
programming documentation for the Q family, it appears that Microchip
wanted to make programming the Q family chips more compatible with SPI
interfaces; so for these chips the data is exchanged in 8-bit chunks,
and everything is MSB-first.

The older chips, on the other hand, use a curious custom protocol. Each
"packet" exchanged with the target consists of a 4-bit command
concatenated with a 16-bit payload. Everything is LSB first, and a few
programming commands require that the clock be held high, during the 4th
bit of the command, to externally time the programming process. So
custom bitbanging firmware on the programming "pod" (historically, in my
case, an S08 microcontroller) is essentially a necessity.

Here is the sad part of the story: After writing a bunch of new code for
the STM32 and spending a few days trying everything I could think of,
I was unsuccessful in getting either of the two 14K50 chips currently in
my possession - both of which I successfully programmed in 2014 - to
enter programming mode!

After failing with the STM32, I decided to revisit the S08 code. After
making a few changes to it - muforth has moved on in some ways since
2014 - I programmed the new firmware onto an S08, hoping that, since it
worked before, it would work again. Sadly, I had the same result:
nothing I did would convince the 14K50 to go into programming mode.

I'm giving up for now on this, but since this code is potentially usable
for programming a huge variety of PIC18F chips, I will probably revisit
it.

For now I'm checking in these changes, which include some
simplifications to the half-duplex "SPI" bitbang code that I'm currently
using - with great success! - to program the PIC18F-Q family.