[pic18] Not-quite-working chat peer code for PIC18
I wanted to check this is in before trying some approaches to fixing it.

First, why a chat peer, and what is it? This is code that runs on the
host and emulates the behavior of a target's serial chat code. However,
instead of communicating over a hardware serial interface it uses
a pseudo-tty (pty).

I have written chat peers for the S08 and MSP430, and they can be very
useful for debugging, for testing out flash programming code, etc.

The problem with this code: I was storing and manipulating the target
memory addresses - the equivalent of TBLPTR and NVMADR on the device
- as already-converted *host* addresses (ie, they point to a byte in
a host-resident memory image, instead of being a target offset), but on
a machine like the PIC18, with multiple *images* in the flash address
space, it's not easy to map from a host address back to the offset and
image that it corresponds to, which I attempt to do in this code.