[pic18] Added an experimental 24-bit Forth kernel!
Well, the beginnings of one, anyway.

I need to make some changes to the meta-compiler and to the common/stack
code to get this to work, but I'm liking the idea a lot.

An 8-bit machine can emulate an 8n-bit CPU; why not 24 bits?

On the PIC18 this aligns nicely both with the PC and program memory
pointer sizes, and there are several peripherals that use 24 bit
registers. Also, having a bit of "headroom" for computations, above what
you get with 16-bits, would be nice!