[several] Target stack copy code more configurable
I have been experimenting with a 24-bit kernel - meaning, really, that
the stack values are 24-bits - on an 8-bit PIC18 microcontoller.

Previously, the stack copying code used the target's standard "size"
words - \m cell etc - and standard image read and write (image-@ and
image-!).

But with this 24-bit experiment, the size of items on the stack, and the
words to read and write them, are now *different* from the standard
memory words, and need separately configurable implementations.

In the process of making this change, I also renamed the words
image-push and image-pop (which were confusing because image-push was
*reading* the target stack image and *pushing* values to the host stack,
and image-pop doing the reverse); now the words are called item> and
item<, and these correspond nicely to stack> and stack< .

Also, unlike before, where image-@ and image-! did an image+ for every
item, now the main loop calls image+ once, and item> and item< expect
*host* rather than *target* addresses. I think it's overall an
improvement and a simplification.