[s08] Commit region considered harmful!
As I was testing changes to the MSP430 code - changing it to use the
target/common/flash.mu4 "framework" - I realized that having a commit
region that represents the first 4 bytes of the boot code simply doesn't
work. The main problem: the boot region - which immediately follows the
commit - starts on an oddball address that is, importantly, *not* on
a chunk boundary. This completely throws off the chunking - every chunk
afterwards also starts offset by 4 bytes; which means that no chunk is
going to trigger the page erasure necessary for successful flashing.

I removed the commit region from the S08 code, and I have removed it
from the (upcoming) changes to the MSP430 flashing code.

The approach that seemed to work is actually simply a reworking of code
I already had in the MSP430 flashing code, but only just realized was
here. To wit, this:

: ?skip-commit  ( a u - a u)  over @boot =  if  4 -4 v+  then ;

: verify-chunk  ( diff 'target len - diff' 'target+len)
   ?skip-commit
   noisy-verify? @ if
      cr ." verifying "  over u.  dup u.  ." : "
   then
   2dup + push
   ( diff 'target len)  setup-comparison
   \m aligned  \m cell/ for  verify-cell  \m cell+  next  drop
   pop ;

Instead of calling ?skip-commit for every region, and checking in
skip-commit if we are in the fact in the boot code - instead we now do
this:

: skip  ( offset)   'origin +! ;    ( change region's origin pointer)

-: verify-1  save-trims
     flash  verify-region  report
      boot  4 skip  ( don't include commit in verify)
            verify-region  report  or
            -4 skip
     trims  verify-region  report  or
   vectors  verify-region  report  or ;  is verify-regions

We skip forward before, and backward after, verifying the boot region.

Works like a charm!