[msp430] Chat dummy works with new interact
I changed the interact code a while back to use the "modern" t.Run and
t.Status interfaces.

Chat dummy can now be used to test flashing and even whether chat it
running in ram or not.

After loading the dummy chat code, eg, like this:

  ./muforth -d g2553 -d dummy -f target/MSP430/load-chat.mu4

then try

  chat
  flash-image

You should get a warning that you need to be running the chat code from
RAM.

To do this, pretend to run code in ram, like this:

  @ram 0 t.Run

This will set the pc to @ram. Then re-chat:

  chat

You should now see the message "(running in RAM)" and flash-image will
now work.