[s08,msp430] Fixed flashing of boot region (again!)
Fixes that aren't fixes!

My previous "fix" worked as long as the boot region contained at 4 bytes
of content, but if it was empty (which it often will be) the attempt to
skip the commit will cause the origin address to be 4 bytes *beyond* the
here address; the region size will therefore be negative - or perhaps
a very large unsigned number. Verify basically hangs trying to compare
an infinite amount of memory.

This "fix", which is hopefully finally correct - only does the skip if
the boot region contains at least 4 bytes.