[msp430] Fixed chat-peer
The chat5 flashing code didn't work, mostly because it used a different
memory pointer (mp rather than m) than the chat2 and chat4 code.

I refactored everything, so that all the common code is at the start of
the file; now all chat protocols work correctly, and each one has less
work to do, since all it has to implement is what is *distinct* to that
protocol.