[riscv] Reorg of RISC-V device files
target/RISC-V directly contains only vendor-independent RISC-V files;
below that, things are organized as follows:

  family/chip/board/

or, if the chip is a one-off and it would pointless to declare a chip
family:

  chip/board/

In the case of the GigaDevice gd32vf103 I wasn't sure whether to make it
a one-off or not. I decided to include a family directory, in case GD
releases new chips that I want to support!

The target/RISC-V/board/ directory now contains *symlinks* to the board
load files, which now live underneath the chip that they sport/support.

For example, for SiFive Hifive1 board, the directories nest like this:

  fe310/                | family
    g000/               | chip
      hifive1.mu4       | board load file
      hifive1/          | board support files

and for the Sipeed Longan Nano:

  gd32v/                | family
    f103cbt6/           | chip
      longan-nano.mu4   | board load file
      longan-nano/      | board support files (eg, for the SPI display)

target/ARM/ is organized in a similar way, but not quite as carefully.
I may revisit it and make similar changes there!