[riscv] Connect to your ESP32-C3 and look around!
I connected the uart-bootloader to the "interact" code, which means that
du (dumping of memory) and dis (disassembly) both work.

Here is how to do it:

(1) Figure out which USB-serial device corresponds to your ESP32-C3
board. Plug it in and see. (This is very OS-specific. Some details can
be found at https://muforth.dev/using-usb-serial-adapters/ )

(2) Make muforth/mu/serial-target a symlink pointing to the device you
discovered in (1).

(3) Load muforth like this:

  ./muforth -f target/RISC-V/board/esp32-c3-devkitc-02

Right now the only "board" definition is for the above devkit, but the
DevKitM-01 should also work (and possibly other ESP32-C3-based boards).

After the code loads, type

  boot

muforth will seem to hang; it is waiting for the board. Now press and
hold the boot (B) button; press and release the reset (R) button; and
release the boot button.

You should see the ROM boot message, a message from muforth that says
"Received ROM header". At this point muforth will send two SYNC packets
to the bootloader and check the response. If all has gone well, you can
do exploring. An interesting thing to try is to take a look at the jump
table at the start of ROM 0:

  @rom0 dis

The first thing we see is a jump. Let's follow it and see where it goes
(but remember where we came from). Press "c" to call, then press space
or return several times. We see code setting various fundamental things,
including a write to mtvec (the pointer to the interrupt/exception
vector table).

Press "r" to return to where we were, then press space or return several
times. There are some 0 cells that disassemble as illegal instructions,
but starting at 4000_0018 we see a series of jump instructions.

This is one of the ROM jump tables! The destinations of each of these
jumps is defined here:

  https://github.com/espressif/esp-idf/blob/v5.5/components/esp_rom/esp32c3/ld/esp32c3.rom.ld

There is something else interesting in these first few words of the ROM,
though. This file:

  https://github.com/espressif/esp-idf/blob/v5.5/components/esp_rom/esp32c3/ld/esp32c3.rom.version.ld

tells us that the chip ID is at 4000_0010, and the ECO version (I'm
still unclear on what that is, but we will find out soon) is at
4000_0014.

Try this:

  @rom0 du

and press space or return until you can see the values at 0010 and 0014.

On my chip (from 2021) I see this:

40000010 >  05 00 00 00  03 00 00 00  6f 80 74 27  6f 80 b4 20
         >     00000005     00000003     2774806f     20b4806f

My chip ID is 5 and my ECO version is 3.

There are more .ld files in the components/esp_rom/esp32c3/ld/ directory
that we looked at above with "eco3" and "eco7" in their names, so it
suggests a place to look further. (At firsts glance it seems to have to
do with the bluetooth radio.)

There you have it. The first steps of a long(er) journey!