[riscv] Made connecting to ESP32-C3 bootloader easier
In particular, easier to *re*connect.

When you first try to connect, use

  hi

This will prompt you to press the Boot and Reset buttons in a particular
sequence, forcing the chip into the bootloader; muforth will then
execute the "boot" command, which connects the bootloader memory read to
du and dis (see previous commit).

If you have reloaded muforth but not reset the chip, it will still be in
the bootloader, so there is no need to say "hi"; instead just type

  boot

and you can now go exploring!