[riscv] Programmatically put ESP32-C3-devkit into bootloader mode!
Not sure if this works for other C3 boards; the DevKitM-01 and
DevKitC-02 have a strange way of cross-wiring the two signals (almost
like a flip-flop, using transistors), but my technique seems to work.

Boards like the seeed Xiao-ESP32-C3 - which use the internal
USB-serial/JTAG interface instead of an external USB-serial chip - may
or may not work with the current technique.