[core] Stack display code fixed (broken since 2018!!)
The stack display code - which shows as many of the values on the stack
as will fit on a single line - is turned on by default. It is an
incredibly useful feedback tool; after every Ok prompt, the contents of
the stack are printed.

When I first added this feature, I was printing the stack in hex -
regardless of the setting of radix. This was a useful default while
getting things and debugging, but it's not a very user-friendly
interface choice. It is prefereable to honor the user's radix choice
wherever possible.

The stack display, however, used fixed-sized fields that could perfectly
fit a 64-bit number, as long as it was displayed in *hex*. When
I changed the code to display numbers in the current radix, two things
broke:

  * The number of items to display was calculated based on the
    fixed-size field, and so would be wrong for non-hex radices;

  * Large numbers - taking more than 16 characters to display - would
    run into their neighbors, making the stack display illegible.

There was another user interface problem with the existing code: numbers
are, by default, printed with separators: for decimal, a comma for every
3 digits; for every other radix, an underline every 4.

The stack display, however, didn't do this. For long numbers it was
impossible to visually parse the number without printing it out.

I wanted to fix all of these problems - and they are now all fixed!

Stack display:

  * displays numbers in the current radix;

  * uses variable-sized, rather than fixed, fields, adding three spaces
    between numbers;

  * only displays as many values as will fit on the screen, and is very
    careful about this;

  * allows turning separators on (+sep) and off (-sep), like the regular
    number printing words: . and u.

If you want to see the new code in action, try this:

  ./muforth
  1 2 3

You should see the stack, and an indicator that there are 3 items on it.

  binary
  -1 u2/

Now there is a very large binary number (63 digits plus separators) on
the top of the stack.

  over

probably didn't change much

  over

Depending on wide your text display is, the number of displayed values
has now likely changed.

  -sep

might show more value, because they take up less space

  +sep hex

again, will reformat the space based on the radix.

You get the idea!

It's much better than the way things have been for the last ~6 years!