[common] Fixed stack<
The number of items read from the target's stack is "clamped" so that if
the target's stack pointer gets corrupted the host doesn't try to read a
huge chunk of memory and possibly crash.

Unfortunately, if the target stack depth exceeded the limit (which was
set at 12 items; I raised it to 32), instead of copying the *top* twelve
items, it copied the bottom twelve!