[pic18] Chat no longer clobbers TBLPTR/PGMADR when running code
I'm doing some experiments that involve interactively using TBLPTR
(which I have renamed PMGADR for various reasons ;-) and I was mystified
that it kept getting clobbered whenever I executed code interactively.

It turns out I was using it in chat as a "holding place" for the address
of the code to execute. The address would first be written to TBLPTR and
then, when chat received the "run" command, copied into
PCLATU/PCLATH/PCL, which would cause a jump to the code.

Instead of first writing TBLPTR and then copying over, now we instead
write the code address *directly* to PCLAT/PCL.

Interestingly, after this change the chat code is the same size.