[pic18] Proof-of-concept of using DMA to buffer UART RX
I want to keep the chat code a *polled* implementation, but I am also
thinking of adding a simple multitasker.

Normally with a tasking system you want the receive side of the UART to
be buffered in some kind of "ring" buffer. In the past I have done by
having an interrupt read the byte and write it into the buffer; but
I would ideally like to be able to add "app" code to the device that is
separate from the "chat" code, and have them both act as tasks.

Unfortunately, this creates a conflict: who is responsible for the
contents of the interrupt vector table?

Ideally I would like to let the app use the entire table, but this means
that chat needs to remain a *polling* implementation.

In order to remain polled but also resilient to delays, what else can we
use to buffer the received bytes?

The pic18-q devices have 6 DMA instances, and these are perfectly suited
to this task.

This task tests out the idea; it seems to work perfectly.

I am now thinking about further tests and then how to integrate it into
the Q-family chat code.